equities_data_service.py: - delete_equity() raise exception if account id is not found - fix equities_as_dict() - add an accountIds list filter for return set - add query_values_over_time_by_accounts() - add balance_over_time_data() - add balance_sheet_over_time()
ValueSnapshot.py:
- change year to int field
- add recorded_date as a field

balancesheet.enums.py:
- abstract @classmethods to enums.py

equityManager.py
- Add notification if equity entered successfully
- raise exception if equity not found when deleting
- Add notificaiton to user if deleted successfully
- add print_balance_sheet()
- add plot_balance_over_time()

account_data_service.py:
- change query_account() to accept a list of accountNames

ledger_data_service.py:
- change parameters on create to their respective enum types
- update query_ledger() to accept a list of account names

accountManager.py:
- general formatting

ledgerkeeper.enums.py:
- abstract @classmethods to enums.py

ledgerManager.py:
- update enum comparisons based on _request_enum() return

cliInteractionManager.py:
- update request_enum() to return the enum itself rather than the name
- add compatibility in pretty_print_items() to accept a dataFrame
- remove old comments

CliRunner.py:
- update main loop to correctly return if user selects to go back
- add input for ledger_sub_menu()
- add print_sub_menu_switch()
- add print_sub_menu()
- add print_sub_menu_loop()
- add return input for equity_sub_menu()
- add PlotType.ASSET_LIABILITY_NET_OVER_TIME to plot_switch()
- add plot_sub_menu()
- add plot_request_loop()
- add print_report()

enums.py:
- add MyEnum() class containing @classmethod logic for enums

plotter.py:
- move out of ledgerKeeper.py
- add account filter for plot_history_by_category()
- add plot_assets_liabilities_worth_over_time()

+++ balancesheet/mongoData/equities_data_service.py
@@ -2,18 +2,21 @@ from balancesheet.mongoData.equities import Equity
 from balancesheet.mongoData.valueSnapshot import ValueSnapshot
 from typing import List, Dict
 from balancesheet.enums import AssetType, LiabiltyType, EquityContingency, EquityTimeHorizon, EquityClass, EquityStatus
-
+from mongoengine.queryset.visitor import Q
+import pandas as pd
+import datetime
+from dateutil.relativedelta import relativedelta
 
 
 def enter_equity(name:str,
-    description:str,
-    accountId: str,
-    equityClass: EquityClass,
-    equityType: [AssetType, LiabiltyType],
-    equityTimeHorizon : EquityTimeHorizon,
-    equityStatus:EquityStatus,
-    equityContingency:EquityContingency,
-    interestRate: float = 0.0) -> Equity:
+                 description:str,
+                 accountId: str,
+                 equityClass: EquityClass,
+                 equityType: [AssetType, LiabiltyType],
+                 equityTimeHorizon : EquityTimeHorizon,
+                 equityStatus:EquityStatus,
+                 equityContingency:EquityContingency,
+                 interestRate: float = 0.0) -> Equity:
 
     equity = Equity()
     equity.name = name
@@ -31,38 +34,70 @@ def enter_equity(name:str,
     return equity
 
 def enter_if_not_exists(name:str,
-    description:str,
-    accountId:str,
-    equityClass: EquityClass,
-    equityType: [AssetType, LiabiltyType],
-    equityTimeHorizon : EquityTimeHorizon,
-    equityStatus:EquityStatus,
-    equityContingency:EquityContingency,
-    interestRate: float = 0.0) -> Equity:
+                        description:str,
+                        accountId:str,
+                        equityClass: EquityClass,
+                        equityType: [AssetType, LiabiltyType],
+                        equityTimeHorizon : EquityTimeHorizon,
+                        equityStatus:EquityStatus,
+                        equityContingency:EquityContingency,
+                        interestRate: float = 0.0) -> Equity:
 
     existing = Equity.objects(name=name)
 
     if len(existing) > 0:
         return None
     else:
-        return enter_equity(name,description, accountId, equityClass, equityType, equityTimeHorizon, equityStatus, equityContingency, interestRate)
+        return enter_equity(name, description, accountId, equityClass, equityType, equityTimeHorizon, equityStatus,
+                            equityContingency, interestRate)
 
 def delete_equity(account_id:str, equity_name:str):
     equity = equity_by_account_and_name(account_id, equity_name)
-
+    if equity is None:
+        raise Exception(f"Account Id: [{account_id}], equity_name {equity_name} not found.")
     equity.delete()
 
 
-def equities_as_dict() -> Dict[int, str]:
-    accounts = query_equities('')
-    return {i + 1: accounts[i].account_name for i in range(0, len(accounts))}
+def equities_as_dict(accountIds:List[str] = None) -> Dict[int, str]:
 
-def query_equities(query) -> List[Equity]:
+    equities = query_equities('', raw_return=False, accountIds=accountIds)
+    return {i + 1: equities[i].name for i in range(0, len(equities))}
+
+def query_equities(query, raw_return=True, accountIds:List[str] = None) -> List[Equity]:
     if query == "":
-        return Equity.objects().order_by('name')
+        equities = Equity.objects()
     else:
-        return Equity.obects(__raw__=query).order_by('name')
+        equities = Equity.objects().filter(__raw__=query)
 
+    if accountIds is not None:
+        account_query = Q(account_id__in=accountIds).to_query(Equity)
+        equities = equities.filter(__raw__=account_query)
+
+    equities = equities.order_by('name')
+
+    if raw_return:
+        return equities.as_pymongo()
+    else:
+        ret_equities = []
+        for l in equities.order_by('name'):
+            ret_equities.append(l)
+
+        return ret_equities
+
+def query_values_over_time_by_accounts(query, raw_return=True, accountIds:List[str] = None):
+    equities = query_equities("", raw_return=False, accountIds=accountIds)
+
+    values =[]
+    for equity in equities:
+        for value in equity.snapshots:
+            values.append({"accountId":equity.account_id, "name": equity.name, "class": equity.equityClass,
+                           "year": value.year, "month": value.month, "value": value.value})
+
+    return values
+
+    if accountIds is not None:
+        account_query = Q(account_id__in=accountIds).to_query(Equity)
+        equities = equities.filter(__raw__=account_query)
 
 
 
@@ -87,7 +122,7 @@ def record_value_on_equity(equity: Equity,
     valueSnapshot.month = month
     valueSnapshot.value = value
 
-    equity = Equity.objects(id=equity.id)
+    equity = Equity.objects(id=equity.id).first()
     updated = False
     for snap in equity.snapshots:
         if (snap.year == year) and (snap.month == month):
@@ -101,8 +136,33 @@ def record_value_on_equity(equity: Equity,
 
     return valueSnapshot
 
+def balance_over_time_data(relevant_months: int, accountIds: List[str] = None):
+    if accountIds is None:
+        values = query_values_over_time_by_accounts("", True)
+    else:
+        values = query_values_over_time_by_accounts("", True, accountIds)
+
+    data = pd.DataFrame(values)
+
+    start_date = datetime.datetime.now() - relativedelta(months=+relevant_months)
+    month = start_date.month
+    year = start_date.year
+    timed_data = data[(12 * data.year + data.month) >= 12 * year + month]
+    timed_data['BOM'] = pd.to_datetime(dict(year=timed_data.year, month=timed_data.month, day=1))
+
+    return timed_data
 
 
+def balance_sheet_over_time(relevant_months: int, accountIds: List[str] = None):
+    timed_data = balance_over_time_data(relevant_months=relevant_months, accountIds=accountIds)
+
+    pivot = pd.pivot_table(timed_data, values=['value'], index=['accountId', 'class', 'name'], columns=['BOM']).fillna(0)
+    pivot.sort_index( axis=1, ascending=False, inplace=True)
+    pivot.sort_values(by=['accountId', 'class', 'name'], inplace=True)
+
+    # with pd.option_context('display.max_rows', 500, 'display.max_columns', 2000, 'display.width', 250):
+    #     print(pivot)
+    return pivot
 
 if __name__ == "__main__":
     import mongo_setup
